Remove create template xservice
